+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Tests:
   * Added [version-related unit tests](./tests/test_bme_version.sh).
   * Output [support functions](./tests/maketests.sh).
+* **python3-virtualenvs** module: new environment variable *'BME_PYTHON_VERSION'* allows configuration for the Python binary used to build virtualenvs.
 
 ## v1.7.1 (2024-FEB-04)
 * differences from [previous tag](/../../compare/v1.7.0…v1.7.1).